requirements.txt
django
django-rest-framework
psycopg2
dj_database_url
